requirements.txt
Flask=2.0.3
requests=2.27.1
python-binance=1.0.15
pandas=1.4.
gunicorn